Time Complexity
A small frog wants to get to the other side of the road. The frog is currently located at position X and wants to get to a position greater than or equal to Y. The small frog always jumps a fixed distance, D. Count the minimal number of jumps that the small frog must perform to reach its target.
O(1) - this program will check if the distance which the frog needs to travel (Y-X) is mod D meaning it would divide perfectly, if it does not then do the divide and add 1.