Time Complexity - O(N)
This program takes a non empty array A, splits it in two and finds the absolute minimal difference that can be achieved between the two parts and return the lowest number.
O(N) - first we assign the first possible value to min, if the split was the first indicies minus the rest. Then check the other possible values for min in the second for loop, comparing both to see which is smaller.